+++ test-performance/src/androidTest/java/com/layer/xdk/test/performance/PerformanceTestComponent.java
@@ -1,0 +1,15 @@
+package com.layer.xdk.test.performance;
+
+import com.layer.xdk.ui.DefaultXdkUiModule;
+import com.layer.xdk.ui.conversation.adapter.ConversationItemsAdapter;
+
+import javax.inject.Singleton;
+
+import dagger.Component;
+
+@Singleton
+@Component(modules = DefaultXdkUiModule.class)
+public interface PerformanceTestComponent {
+
+    ConversationItemsAdapter conversationItemsAdapter();
+}