+++ ui/src/test/java/com/layer/xdk/ui/message/model/MessageModelTest.java
@@ -12,11 +12,11 @@ import android.test.mock.MockContext;
 import com.layer.sdk.LayerClient;
 import com.layer.sdk.messaging.Message;
 import com.layer.sdk.messaging.MessagePart;
-import com.layer.xdk.ui.stub.ConversationStub;
-import com.layer.xdk.ui.stub.IdentityStub;
-import com.layer.xdk.ui.stub.LayerClientStub;
-import com.layer.xdk.ui.stub.MessagePartStub;
-import com.layer.xdk.ui.stub.MessageStub;
+import com.layer.xdk.test.common.stub.ConversationStub;
+import com.layer.xdk.test.common.stub.IdentityStub;
+import com.layer.xdk.test.common.stub.LayerClientStub;
+import com.layer.xdk.test.common.stub.MessagePartStub;
+import com.layer.xdk.test.common.stub.MessageStub;
 
 import org.junit.BeforeClass;
 import org.junit.Test;